Update composer.json
This is the format we will actually be using.
For now we'll keep the author as the phpBB team, we can modify this later to include 
the chameleon team if we feel that's appropriate.
As this isn't actually the 3.2.0 style yet and we should be using branch aliases anyway 
I've updated that. phpBB we require self-version (3.2.x) for obvious reasons although this means
Chameleon versions will be the same as phpBB version, this will look odd but makes life so 
much easier in the long run. We don't have a requirement of a specific PHP version.

In the future we might look to add twig as requirement for styles so they can use custom twig versions
this will cause conflicts with multiple styles requiring different versions so to do that we'd need to properly
educate style authors about that requirement.