Refactor spacing and establish preset scale
I have gone back and forth on the best way to handle spacing.
I finaly come across a really useful point that hits home.
https://medium.com/tap-to-dismiss/constraint-layout-for-designers-3c665cb4d074
Outside of componets and possibly objects all spacing should be static,
that is to say preset variables to easily sync design and code. This is will be
handled different for components and possibly objects to preserve easy
customization they wil be dynamicly spaced accoriding the there own scope
but still adhere to the same established spacing scale.